+++ ChangeLog.md
@@ -5,7 +5,7 @@ v0.6.0 (10Sep20)
 -----------------------
 
 - Numerical computation of the Hessian matrix 
-- Calculator for polydisperse systems 
+- Calculator for polydisperse systems in which particles interact via a pair potential 
 - Analytic computation of Hessian for polydisperse systems
 - Bug fix in tests eam_calculator_forces_hessian